+++ Pasteleria/src/main/java/com/example/pasteleria/repositories/ReviewRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReviewRepository extends CrudRepository<ReviewModel,Integer> {
+public interface ReviewRepository extends CrudRepository<ReviewModel,Long> {
 
 }
